remove all cache
