--- src/Pages/CategoryPage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function CategoryPage() {
-  return <div>CategoryPage</div>;
-}
-
-export default CategoryPage;
