--- src/Pages/SearchPage.tsx
@@ -1,13 +1,0 @@
-import { PageContainer } from 'components/common/CommonComponents';
-import SearchField from 'components/search/SearchField';
-import React from 'react';
-
-function SearchPage() {
-  return (
-    <PageContainer>
-      <SearchField />
-    </PageContainer>
-  );
-}
-
-export default SearchPage;
